set alarm on watch failures
Prior to this commit we had no way of notifying other portions of the
system when we have a problem reading a configuration source. This
commit uses the `sasl` `alarm_handler` to set an alarm of the following
format `{:vapor, {plan_layer, plan_provider}}`. We also redact the
alarm `reason` because passing through lower-level `reason`s from their
error tuple could expose secrets to logging.